Refactor modules (#674)
* separate logging functions to own module

* refactor update

* fix circular import

* updates

* update

* version bump
