+++ addon.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <addon id="plugin.video.themoviedb.helper"
-version="4.6.24"
+version="4.6.25"
 name="TheMovieDb Helper"
 provider-name="jurialmunkey">
 <requires>
