+++ resources/lib/addon/plugin.py
@@ -1,10 +1,8 @@
 import re
 import xbmc
-import hashlib
-import traceback
-from xbmcgui import Dialog
 from xbmcaddon import Addon as KodiAddon
 from resources.lib.addon.constants import LANGUAGES
+""" Top level module only import constants """
 
 
 ADDON = KodiAddon('plugin.video.themoviedb.helper')
@@ -23,8 +21,7 @@ ADDONSETSETTINGROUTE = {
     'str': ADDON.setSettingString
 }
 
-_addonlogname = '[plugin.video.themoviedb.helper]\n'
-_debuglogging = ADDON.getSettingBool('debug_logging')
+
 _executebuiltin = xbmc.executebuiltin
 _getcondvisibility = xbmc.getCondVisibility
 _getinfolabel = xbmc.getInfoLabel
@@ -101,40 +98,6 @@ def reconfigure_legacy_params(**kwargs):
 def set_kwargattr(obj, kwargs):
     for k, v in kwargs.items():
         setattr(obj, k, v)
-
-
-def md5hash(value):
-    value = str(value).encode()
-    return hashlib.md5(value).hexdigest()
-
-
-def kodi_log(value, level=0):
-    try:
-        if isinstance(value, list):
-            value = ''.join(map(str, value))
-        if isinstance(value, bytes):
-            value = value.decode('utf-8')
-        logvalue = f'{_addonlogname}{value}'
-        if level == 2 and _debuglogging:
-            xbmc.log(logvalue, level=xbmc.LOGINFO)
-        elif level == 1:
-            xbmc.log(logvalue, level=xbmc.LOGINFO)
-        else:
-            xbmc.log(logvalue, level=xbmc.LOGDEBUG)
-    except Exception as exc:
-        xbmc.log(f'Logging Error: {exc}', level=xbmc.LOGINFO)
-
-
-def kodi_traceback(exception, log_msg=None, notification=True, log_level=1):
-    if notification:
-        head = f'TheMovieDb Helper {get_localized(257)}'
-        Dialog().notification(head, get_localized(2104))
-    msg = f'Error Type: {type(exception).__name__}\nError Contents: {exception.args!r}'
-    msg = [log_msg, '\n', msg, '\n'] if log_msg else [msg, '\n']
-    try:
-        kodi_log(msg + traceback.format_tb(exception.__traceback__), log_level)
-    except Exception as exc:
-        kodi_log(f'ERROR WITH TRACEBACK!\n{exc}\n{msg}', log_level)
 
 
 def get_language():
