+++ resources/lib/api/trakt/api.py
@@ -3,7 +3,7 @@ from xbmc import Monitor
 from xbmcgui import Dialog
 from json import loads, dumps
 from resources.lib.addon.window import get_property
-from resources.lib.addon.plugin import kodi_log, get_localized, get_setting, set_setting
+from resources.lib.addon.plugin import get_localized, get_setting, set_setting
 from resources.lib.addon.parser import try_int
 from resources.lib.addon.timedate import set_timestamp, get_timestamp
 from resources.lib.files.cache import CACHE_SHORT, CACHE_LONG, use_simple_cache
@@ -12,7 +12,7 @@ from resources.lib.api.request import RequestAPI
 from resources.lib.api.trakt.items import TraktItems
 from resources.lib.api.trakt.decorators import is_authorized, use_activity_cache
 from resources.lib.api.trakt.progress import _TraktProgress
-# from resources.lib.addon.decorators import timer_report
+from resources.lib.addon.logger import kodi_log
 
 
 API_URL = 'https://api.trakt.tv/'
@@ -303,7 +303,6 @@ class _TraktSync():
             return activities.get(activity_type, {})
         return activities.get(activity_type, {}).get(activity_key)
 
-    # @timer_report('_get_last_activity')
     @is_authorized
     def _get_last_activity(self, activity_type=None, activity_key=None, cache_refresh=False):
         if not self.last_activities:
