+++ resources/lib/files/cache.py
@@ -1,10 +1,8 @@
-from resources.lib.addon.plugin import kodi_log, format_name
-from resources.lib.addon.decorators import try_except_log
+from resources.lib.addon.plugin import format_name
 from resources.lib.files.simplecache import SimpleCache
 from resources.lib.files.utils import get_pickle_name
-# from resources.lib.files.utils import pickle_deepcopy
-# from threading import Thread
-# from resources.lib.addon.decorators import TimerList
+from resources.lib.addon.logger import kodi_log, try_except_log
+
 
 CACHE_LONG = 14
 CACHE_SHORT = 1
