+++ resources/lib/files/simplecache.py
@@ -12,7 +12,7 @@ import sqlite3
 from xbmcgui import Window
 from xbmc import Monitor, sleep
 from contextlib import contextmanager
-from resources.lib.addon.plugin import kodi_log
+from resources.lib.addon.logger import kodi_log
 from resources.lib.addon.timedate import set_timestamp
 from resources.lib.files.utils import get_file_path
 from resources.lib.files.utils import json_loads as data_loads
