+++ resources/lib/player/players.py
@@ -4,16 +4,17 @@ from xbmcgui import Dialog
 from xbmcplugin import setResolvedUrl
 from xbmcaddon import Addon as KodiAddon
 from resources.lib.addon.window import get_property
-from resources.lib.addon.plugin import ADDONPATH, PLUGINPATH, format_folderpath, kodi_log, get_localized, get_setting, executebuiltin, get_infolabel
+from resources.lib.addon.plugin import ADDONPATH, PLUGINPATH, format_folderpath, get_localized, get_setting, executebuiltin, get_infolabel
 from resources.lib.addon.parser import try_int, try_float
 from resources.lib.addon.constants import PLAYERS_PRIORITY
-from resources.lib.addon.decorators import busy_dialog, ProgressDialog
+from resources.lib.addon.dialog import BusyDialog, ProgressDialog
 from resources.lib.items.listitem import ListItem
 from resources.lib.files.utils import read_file, normalise_filesize
 from resources.lib.api.kodi.rpc import get_directory, KodiLibrary
 from resources.lib.player.details import get_item_details, get_detailed_item, get_playerstring, get_language_details
 from resources.lib.player.inputter import KeyboardInputter
 from resources.lib.player.configure import get_players_from_file
+from resources.lib.addon.logger import kodi_log
 
 
 def string_format_map(fmt, d):
@@ -74,7 +75,7 @@ def resolve_to_dummy(handle=None, stop_after=1, delay_wait=0):
         return -1
 
     # Added delay
-    with busy_dialog(False if delay_wait < 1 else True):
+    with BusyDialog(False if delay_wait < 1 else True):
         Monitor().waitForAbort(delay_wait)
 
     # Success
