+++ resources/lib/script/router.py
@@ -4,10 +4,9 @@
 import sys
 import xbmcvfs
 from json import dumps
-from xbmcgui import Dialog
 from resources.lib.addon.window import get_property
-from resources.lib.addon.plugin import reconfigure_legacy_params, kodi_log, format_folderpath, convert_type, get_localized, get_setting, set_setting, executebuiltin, get_infolabel
-from resources.lib.addon.decorators import busy_dialog
+from resources.lib.addon.plugin import reconfigure_legacy_params, format_folderpath, convert_type, get_localized, get_setting, set_setting, executebuiltin, get_infolabel
+from resources.lib.addon.dialog import BusyDialog, ProgressDialog, kodi_dialog_contextmenu, kodi_dialog_yesno, kodi_dialog_ok, kodi_dialog_input, kodi_dialog_textviewer
 from resources.lib.addon.parser import encode_url, parse_paramstring
 from resources.lib.files.downloader import Downloader
 from resources.lib.files.utils import dumps_to_file, validify_filename, read_file
@@ -25,12 +24,13 @@ from resources.lib.player.players import Players
 from resources.lib.player.configure import configure_players
 from resources.lib.monitor.images import ImageFunctions
 from resources.lib.script.sync import sync_trakt_item
+from resources.lib.addon.logger import kodi_log
 
 
 # Get TMDb ID decorator
 def get_tmdb_id(func):
     def wrapper(*args, **kwargs):
-        with busy_dialog():
+        with BusyDialog():
             if not kwargs.get('tmdb_id'):
                 kwargs['tmdb_id'] = TMDb().get_tmdb_id(**kwargs)
         return func(*args, **kwargs)
@@ -60,13 +60,13 @@ def is_in_kwargs(mapping={}):
 
 
 def play_media(**kwargs):
-    with busy_dialog():
+    with BusyDialog():
         kodi_log(['lib.script.router - attempting to play\n', kwargs.get('play_media')], 1)
         executebuiltin(f'PlayMedia({kwargs.get("play_media")})')
 
 
 def run_plugin(**kwargs):
-    with busy_dialog():
+    with BusyDialog():
         kodi_log(['lib.script.router - attempting to play\n', kwargs.get('run_plugin')], 1)
         executebuiltin(f'RunPlugin({kwargs.get("run_plugin")})')
 
@@ -85,18 +85,18 @@ def delete_cache(delete_cache, **kwargs):
         'Item Details': lambda: ItemBuilder()}
     if delete_cache == 'select':
         m = [i for i in d]
-        x = Dialog().contextmenu([get_localized(32387).format(i) for i in m])
+        x = kodi_dialog_contextmenu([get_localized(32387).format(i) for i in m])
         if x == -1:
             return
         delete_cache = m[x]
     z = d.get(delete_cache)
     if not z:
         return
-    if not Dialog().yesno(get_localized(32387).format(delete_cache), get_localized(32388).format(delete_cache)):
+    if not kodi_dialog_yesno(get_localized(32387).format(delete_cache), get_localized(32388).format(delete_cache)):
         return
-    with busy_dialog():
+    with BusyDialog():
         z()._cache.ret_cache()._do_delete()
-    Dialog().ok(get_localized(32387).format(delete_cache), get_localized(32389))
+    kodi_dialog_ok(get_localized(32387).format(delete_cache), get_localized(32389))
 
 
 @map_kwargs({'play': 'tmdb_type'})
@@ -175,7 +175,7 @@ def related_lists(tmdb_id=None, tmdb_type=None, season=None, episode=None, conta
     items = get_basedir_details(tmdb_type=tmdb_type, tmdb_id=tmdb_id, season=season, episode=episode, include_play=include_play)
     if not items or len(items) <= 1:
         return
-    choice = Dialog().contextmenu([i.get('label') for i in items])
+    choice = kodi_dialog_contextmenu([i.get('label') for i in items])
     if choice == -1:
         return
     item = items[choice]
@@ -196,8 +196,8 @@ def related_lists(tmdb_id=None, tmdb_type=None, season=None, episode=None, conta
 
 def update_players():
     players_url = get_setting('players_url', 'str')
-    players_url = Dialog().input(get_localized(32313), defaultt=players_url)
-    if not Dialog().yesno(
+    players_url = kodi_dialog_input(get_localized(32313), defaultt=players_url)
+    if not kodi_dialog_yesno(
             get_localized(32032),
             get_localized(32314).format(players_url)):
         return
@@ -212,11 +212,11 @@ def update_players():
 def refresh_details(tmdb_id=None, tmdb_type=None, season=None, episode=None, confirm=True, **kwargs):
     if not tmdb_id or not tmdb_type:
         return
-    with busy_dialog():
+    with BusyDialog():
         details = ItemBuilder().get_item(tmdb_type, tmdb_id, season, episode, cache_refresh=True) or {}
         details = details.get('listitem')
     if details and confirm:
-        Dialog().ok('TMDbHelper', get_localized(32234).format(tmdb_type, tmdb_id))
+        kodi_dialog_ok('TMDbHelper', get_localized(32234).format(tmdb_type, tmdb_id))
         container_refresh()
     return details
 
@@ -238,14 +238,14 @@ def user_list(user_list, user_slug=None, **kwargs):
 
 
 def delete_list(delete_list, **kwargs):
-    if not Dialog().yesno(get_localized(32358), get_localized(32357).format(delete_list)):
+    if not kodi_dialog_yesno(get_localized(32358), get_localized(32357).format(delete_list)):
         return
     TraktAPI().delete_response('users/me/lists', delete_list)
     container_refresh()
 
 
 def rename_list(rename_list, **kwargs):
-    name = Dialog().input(get_localized(32359))
+    name = kodi_dialog_input(get_localized(32359))
     if not name:
         return
     TraktAPI().post_response('users/me/lists', rename_list, postdata={'name': name}, response_method='put')
@@ -287,7 +287,7 @@ def image_colors(image_colors=None, **kwargs):
 
 def library_update(**kwargs):
     if kwargs.get('force') == 'select':
-        choice = Dialog().yesno(
+        choice = kodi_dialog_yesno(
             get_localized(32391),
             get_localized(32392),
             yeslabel=get_localized(32393),
@@ -303,10 +303,10 @@ def library_update(**kwargs):
 
 
 def log_request(**kwargs):
-    with busy_dialog():
+    with BusyDialog():
         kwargs['response'] = None
         if not kwargs.get('url'):
-            kwargs['url'] = Dialog().input('URL')
+            kwargs['url'] = kodi_dialog_input('URL')
         if not kwargs['url']:
             return
         if kwargs.get('log_request').lower() == 'trakt':
@@ -314,7 +314,7 @@ def log_request(**kwargs):
         else:
             kwargs['response'] = TMDb().get_response_json(kwargs['url'])
         if not kwargs['response']:
-            Dialog().ok(kwargs['log_request'].capitalize(), f'{kwargs["url"]}\nNo Response!')
+            kodi_dialog_ok(kwargs['log_request'].capitalize(), f'{kwargs["url"]}\nNo Response!')
             return
         filename = validify_filename(f'{kwargs["log_request"]}_{kwargs["url"]}.json')
         dumps_to_file(kwargs, 'log_request', filename)
@@ -322,13 +322,13 @@ def log_request(**kwargs):
         msg = (
             f'[B]{kwargs["url"]}[/B]\n\n{xbmcvfs.translatePath("special://profile/addon_data/")}\n'
             f'plugin.video.themoviedb.helper/log_request\n{filename}')
-        Dialog().ok(kwargs['log_request'].capitalize(), msg)
-        Dialog().textviewer(filename, dumps(kwargs['response'], indent=2))
+        kodi_dialog_ok(kwargs['log_request'].capitalize(), msg)
+        kodi_dialog_textviewer(filename, dumps(kwargs['response'], indent=2))
 
 
 def sort_list(**kwargs):
     sort_methods = get_sort_methods() if kwargs['info'] == 'trakt_userlist' else get_sort_methods(True)
-    x = Dialog().contextmenu([i['name'] for i in sort_methods])
+    x = kodi_dialog_contextmenu([i['name'] for i in sort_methods])
     if x == -1:
         return
     for k, v in sort_methods[x]['params'].items():
@@ -337,7 +337,6 @@ def sort_list(**kwargs):
 
 
 def log_jsonrpc(**kwargs):
-    dialog = DialogProgress()
     method = "VideoLibrary.GetMovies"
     params = {
         "properties": ["title", "imdbnumber", "originaltitle", "uniqueid", "year", "file"],
@@ -348,12 +347,11 @@ def log_jsonrpc(**kwargs):
         ]}
     }
     response = get_jsonrpc(method, params) or {}
-    dialog.create('JSON_RPC Results', f'{response}')
-    update = 0
-    while update < 100:
-        update += 10
-        dialog.update(update)
-    dialog.close()
+    with ProgressDialog('JSON_RPC Results', f'{response}', total=100, logging=0) as dialog:
+        update = 0
+        while update < 100:
+            update += 10
+            dialog.update(update)
     kodi_log(['JSONRPC:\n', response], 1)
