+++ resources/lib/window/manager.py
@@ -2,9 +2,11 @@ from xbmc import Monitor
 from xbmcgui import Dialog, Window
 import resources.lib.addon.window as window
 from resources.lib.addon.parser import try_int
-from resources.lib.addon.plugin import kodi_log, get_localized, executebuiltin
-from resources.lib.addon.decorators import busy_dialog
+from resources.lib.addon.plugin import get_localized, executebuiltin
+from resources.lib.addon.dialog import BusyDialog
 from resources.lib.api.tmdb.api import TMDb
+from resources.lib.addon.logger import kodi_log
+
 
 PREFIX_PATH = 'Path.'
 PREFIX_QUERY = 'Query'
@@ -219,7 +221,7 @@ class WindowManager(_EventLoop):
             if x == -1:
                 return
             query = split_str[x]
-        with busy_dialog():
+        with BusyDialog():
             tmdb_id = TMDb().get_tmdb_id_from_query(tmdb_type, query, header=query, use_details=True, auto_single=True)
         if not tmdb_id:
             Dialog().notification('TMDbHelper', get_localized(32310).format(query))
